move manhwa18 from legacy
2 websites (.net and .com)
