+++ web/src/engine/websites/_index.ts
@@ -267,6 +267,8 @@ export { default as ManhuaPlus } from './ManhuaPlus';
 export { default as ManhuaScan } from './ManhuaScan';
 export { default as Manhuass } from './Manhuass';
 export { default as ManhuaUs } from './ManhuaUs';
+export { default as Manhwa18 } from './Manhwa18';
+export { default as ManhwaEighteen } from './ManhwaEighteen';
 export { default as ManhwaFull } from './ManhwaFull';
 export { default as ManhwaHentai } from './ManhwaHentai';
 export { default as ManhwaHentaiMe } from './ManhwaHentaiMe';
@@ -691,8 +693,6 @@ export { default as ManhuaID } from './legacy/ManhuaID';
 export { default as ManHuaNiu } from './legacy/ManHuaNiu';
 export { default as ManhuaTai } from './legacy/ManhuaTai';
 export { default as ManhuaZ } from './legacy/ManhuaZ';
-export { default as Manhwa18 } from './legacy/Manhwa18';
-export { default as ManhwaEighteen } from './legacy/ManhwaEighteen';
 export { default as ManhwasMen } from './legacy/ManhwasMen';
 export { default as ManhwaSmut } from './legacy/ManhwaSmut';
 export { default as ManmanApp } from './legacy/ManmanApp';
